requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
gtfs-realtime-bindings==1.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.23.4
protobuf3-to-dict==0.1.5
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.4
Werkzeug==2.3.6